+++ test/libsolidity/ABIJson/empty_name_return_parameters.sol
@@ -1,5 +1,5 @@
 contract test {
-    function f(uint k) public returns (uint) {
+    function f(uint256 k) public returns (uint256) {
         return k;
     }
 }
