+++ test/libsolidity/ABIJson/global_struct.sol
@@ -1,5 +1,5 @@
 pragma abicoder               v2;
-struct S { uint a; }
+struct S { uint256 a; }
 contract C {
     function f(S calldata s) external view {}
     function g(S memory s) public view {}
