+++ test/libsolidity/ABIJson/multiple_methods_order.sol
@@ -1,7 +1,7 @@
 // methods are expected to be in alphabetical order
 contract test {
-    function f(uint a) public returns (uint d) { return a * 7; }
-    function c(uint b) public returns (uint e) { return b * 8; }
+    function f(uint256 a) public returns (uint256 d) { return a * 7; }
+    function c(uint256 b) public returns (uint256 e) { return b * 8; }
 }
 // ----
 //     :test
