+++ test/libsolidity/ABIJson/payable_constructor_abi.sol
@@ -1,5 +1,5 @@
 contract test {
-    constructor(uint param1, test param2, bool param3) payable {}
+    constructor(uint256 param1, test param2, bool param3) payable {}
 }
 // ----
 //     :test
