+++ test/libsolidity/smtCheckerTests/operators/division_4.sol
@@ -1,6 +1,6 @@
 pragma experimental SMTChecker;
 contract C {
-	function f(int x, int y) public pure returns (int) {
+	function f(int256 x, int256 y) public pure returns (int256) {
 		require(y != 0);
 		require(y != -1);
 		return x / y;
